Updating Bank#withdraw()
Currently banking can take a while since it validates that it successfully withdrew the item. I'd like to allow the Script Writer to be able to choose to validate it himself, and instead move on to the next item to withdraw. This will help a lot for bank standing scripts as banking currently takes much too long in it's current form.

Even if we don't want to implement this I would definitely appreciate a future update bringing an option to withdraw without the wait.